Concatenate strings instead of an instance
When providing a filehander to the `RarFile` constructor, `this.rarfile` will not be a path (str), but an instance.
Due the error of concatenating an instance to a string when constructing a `NonRarFile` exception, an `TypeError` is raised instead of the expected `NotRarFile` exception.

Example stacktrace (py.test):
```
self = <rarfile.RarFile object at 0x7fbd02c9d810>

    def _parse_real(self):
        fd = XFile(self.rarfile)
        self._fd = fd
        id = fd.read(len(RAR_ID))
        if id != RAR_ID:
>           raise NotRarFile("Not a Rar archive: "+self.rarfile)
E           TypeError: cannot concatenate 'str' and 'instance' objects

../../.virtualenvs/byte-user-milter/local/lib/python2.7/site-packages/rarfile.py:789: TypeError

```