Give filepath with original separator to unrar cmd
As part of creating a RarInfo object, the library sanitizes the compressed filepaths by standardizing on "/" as a universal path separator. With certain inputs (RAR3 solid archives, and perhaps others as well), this sanitized filename is given as a cmdline argument to the extraction binary. This causes UnRAR.exe to fail to recognize the path, and output no data to stdout, resulting in an empty stream.

Fixed here by giving the original filepath to the unrar cmdline.